Update dubbo-samples-serialization/dubbo-samples-serialization-java sample
