+++ 3-extensions/serialization/dubbo-samples-serialization/dubbo-samples-serialization-java/case-versions.conf
@@ -20,6 +20,6 @@
 # Supported component versions of the test case
 
 # Spring app
-dubbo.version=2.7*, 3.*
-spring.version=4.*, 5.*
-java.version= [>= 8]
+dubbo.version= 3.3.*
+spring.version=6.*
+java.version= [>= 17]
