+++ 3-extensions/serialization/dubbo-samples-serialization/dubbo-samples-serialization-java/dubbo-samples-serialization-java-consumer/src/test/java/DemoServiceIT.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+
+import org.apache.dubbo.config.annotation.DubboReference;
+import org.apache.dubbo.samples.serialization.DemoService;
+import org.apache.dubbo.samples.serialization.DubboConsumer;
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+
+@SpringBootTest(classes = DubboConsumer.class)
+@RunWith(SpringRunner.class)
+public class DemoServiceIT {
+    @DubboReference
+    private DemoService service;
+
+    @Test
+    public void testGreeting() throws Exception {
+        Assert.assertTrue(service.sayHello("dubbo").startsWith("Hello dubbo"));
+    }
+}
